requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
langgraph==0.2.73
langchain-openai>=0.2.0
langchain-core>=0.3.36
openai>=1.0.0
sse-starlette==2.2.1
httpx==0.28.1
aiofiles==24.1.0
PyGithub==2.6.1
pydantic==2.10.6
pydantic-settings==2.7.1
docker==7.1.0
replicate==1.0.7
